Updating existing test for image-modals
